[refactor] Moved the `roundToTwoDecimalPlaces` function and `RoundToTwoDecimalPlacesTest` test to the `core:extensions` module for better code organization
