+++ core/extensions/src/main/java/ir/composenews/extensions/Number.kt
@@ -1,4 +1,4 @@
-package ir.composenews.designsystem.extensionFunction
+package ir.composenews.extensions
 
 fun Number.roundToTwoDecimalPlaces(): String {
     return String.format("%.2f", this)
